requirements.txt
Flask==2.3.3
python-nmap==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2
scapy==2.5.0
paramiko==3.3.1
#ftplib3==0.4.0